feat: extract real config and auth leaf modules (v1.9.1)
Move environment settings and API-key auth out of recon_operator.server
into recon_operator.config and recon_operator.auth. Server re-exports the
same symbols so autonmap patches and package surfaces stay compatible.
